rename json module (not really private)
