+++ setup.cfg
@@ -67,7 +67,7 @@ ignore_missing_imports = True
 [mypy-jsonschema.*]
 ignore_missing_imports = True
 
-[mypy-jupyter_starters._json]
+[mypy-jupyter_starters.json_]
 ignore_missing_imports = True
 ignore_errors = True
