+++ src/jupyter_starters/types.py
@@ -8,6 +8,11 @@ class Status:
     """ pseudo-enum for managing statuses
     """
 
-    DONE = "done"
+    # the starter isn't done yet, and more data is required
     CONTINUING = "continuing"
+
+    # the starter is done, and should not continue
+    DONE = "done"
+
+    # something terrible has happened
     ERROR = "error"
